Doc tweaks
