+++ guide/examples.md
@@ -10,5 +10,5 @@
 | Name        | Source                                | Live demo |
 |:------------|:-------------------------------------:|----------:|
 | To Do List  | [source](./examples/ts/todo-list.ts)  | [Live demo](https://raw.githack.com/MatAtBread/AI-UI/main/guide/examples/ts/ts-example.html?todo-list.ts) |
-| Weather     | [source](./examples/ts/weather-10.ts) | [Live demo](https://raw.githack.com/MatAtBread/AI-UI/main/guide/examples/ts/ts-example.html?weather-10.ts) |
-| Weather JSX | [source](./examples/ts/weather-10.ts) | [Live demo](https://raw.githack.com/MatAtBread/AI-UI/main/guide/examples/ts/ts-example.html?weather-10.ts)
+| Weather     | [source](./examples/ts/weather.ts)    | [Live demo](https://raw.githack.com/MatAtBread/AI-UI/main/guide/examples/ts/ts-example.html?weather.ts) |
+| Weather JSX | [source](./examples/ts/weather.tsx)   | [Live demo](https://raw.githack.com/MatAtBread/AI-UI/main/guide/examples/ts/ts-example.html?weather.tsx)
