+++ guide/examples/ts/ts-example.html
@@ -56,11 +56,11 @@ React >= v17
         "module": 6,
         "target": 7,
         "inlineSourceMap": true,
-        // "jsxFactory": "React.AIUIJSX",
-        // "jsxFragmentFactory": "React.AIUIJSX",
-        // "jsx": "react"
-        "jsx": "react-jsx",
-        "jsxImportSource": "../../../module/esm"
+        "jsxFactory": "React.AIUIJSX",
+        "jsxFragmentFactory": "React.AIUIJSX",
+        "jsx": "react"
+        // "jsx": "react-jsx",
+        // "jsxImportSource": "../../../module/esm"
       }, tsFile);
       const script = document.createElement('script');
       script.setAttribute('type', 'module');
