+++ guide/examples/tsconfig.json
@@ -9,11 +9,11 @@
     "skipLibCheck": true,                                /* Skip type checking all .d.ts files. */
     "strictNullChecks": true,
     "lib": ["DOM","DOM.Iterable","ES2020"],
-    // "jsxFactory": "AIUIJSX",
-    // "jsxFragmentFactory": "AIUIJSX",
-    // "jsx": "react"
-    "jsx": "react-jsx",
-    "jsxImportSource": "../../../module/esm"
+    "jsxFactory": "AIUIJSX",
+    "jsxFragmentFactory": "AIUIJSX",
+    "jsx": "react"
+    // "jsx": "react-jsx",
+    // "jsxImportSource": "../../../module/esm"
 },
   "include": [ "ts/*.ts" ]
 }
